+++ src/opendr/perception/panoptic_segmentation/__init__.py
@@ -1,4 +1,7 @@
-from opendr.perception.panoptic_segmentation.datasets import CityscapesDataset, KittiDataset
+from opendr.perception.panoptic_segmentation.datasets import CityscapesDataset, KittiDataset,\
+    SemanticKittiDataset
 from opendr.perception.panoptic_segmentation.efficient_ps import EfficientPsLearner
+from opendr.perception.panoptic_segmentation.efficient_lps import EfficientLpsLearner
 
-__all__ = ['CityscapesDataset', 'KittiDataset', 'EfficientPsLearner']
+__all__ = ['CityscapesDataset', 'KittiDataset', 'EfficientPsLearner',
+           'SemanticKittiDataset', 'EfficientLpsLearner']