+++ tests/test_clang_format.py
@@ -59,6 +59,7 @@ class TestClangFormat(unittest.TestCase):
         ]
         skippedPaths = [
             'src/opendr/perception/panoptic_segmentation/efficient_ps/algorithm/EfficientPS',
+            'src/opendr/perception/panoptic_segmentation/efficient_lps/algorithm/EfficientLPS',
             'src/opendr/planning/end_to_end_planning/ardupilot',
         ]
         skippedFiles = [